requirements.txt
flask==3.0.3
gunicorn==22.0.0
pillow==10.3.0
pytesseract==0.3.8
